extract tui rendering code to separate module
